--- client/src/App.css
@@ -1,140 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 80px;
-}
-
-.App-header {
-  background-color: #222;
-  height: 150px;
-  padding: 20px;
-  color: white;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
-}
-
-/*-------------------------------------------------------------------
-
-                        1. Preloader
-
--------------------------------------------------------------------*/
-
-.preloader {
-  background: linear-gradient(145deg, rgb(221, 224, 228), rgb(113, 127, 150));
-  position: absolute;
-  z-index: 9999;
-  height: 100%;
-  width: 100%;
-}
-
-.preloader-circle {
-  text-align: center;
-  margin: auto;
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  height: 24px;
-  width: 24px;
-  border: 2px solid rgba(255, 255, 255, .7);
-  border-top: 2px solid rgba(0, 0, 0, .5);
-  border-radius: 50%;
-  animation: rotate 1s infinite linear;
-}
-
-
-
-.App {
-  text-align: center;
-  
-}
-
-.banner {
-  /* background-image: url("./1550080499329.png"); */
-  background-size: auto;
-  background-repeat: no-repeat;
-  height: 800px;
-  width: 100%;
-  margin-top: 0px;
-  overflow: scroll;
-}
-
-.searchBox {
-  
-  border: 1px solid;
-  padding: 4px;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 4px;
-  width: 33%;
-  background-color: white;
-}
-
-.results{
-  background-color: #2196F3;
-  width: 75%;
-  margin-left: auto;
-  margin-right: auto;
-  color: white;
-}
-.book {
-  border: solid 1px black;
-}
-
-h4 {
-  border-bottom: solid 1px white;
-  text-align: center;
-  padding-left: 10px;
-  margin-top: 0px;
-}
-
-p {
-  text-align: left;
-  padding: 10px;
-  background-color: white;
-  margin: 20px;
-  color: black;
-}
-
-.placeholder {
-
-  height: 64px;
-  width: 64px;
-}
-.bookLink {
-  width: 100%;
-  color: white;
-}
-
-/* Discussion classes */
-.post-editor {
-  margin: 1em 20%;
-}
-
-.post-editor-button {
-  float: right;
-  margin-top: 1em;
-}
-
-
-@media(max-width: 900px) {
-  .post-editor {
-    margin: 1em;
-  }
-
-  .post-body {
-    margin: 1em;
-  }
-}
-
