+++ client/src/componenets/NavBar/SignOutButton.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { withFirebase } from '../Firebase';
+
+import SignOutIcon from './SignOut.png';
+
+const SignOutButton = ({ firebase }) => (
+  <div onClick={firebase.doSignOut}>
+    <img className='navIcons' src={SignOutIcon} alt='Sign Out' />
+  </div>
+);
+
+export default withFirebase(SignOutButton);
